rename: MenusIcons.tsx → MenuBarIcons.tsx
